Got a basic link creation working. Needs work.

+++ src/org/pathwayeditor/visualeditor/editingview/IShapePane.java
@@ -41,6 +41,6 @@ public interface IShapePane {
 
 	void removeMouseListener(MouseListener mouseSelectionListener);
 
-	void showPopup(JPopupMenu popup, int x, int y);
+	void showPopup(JPopupMenu popup, double x, double y);
 	
 }
